Switch to an improved layout
Since we're going to build a set of binaries out of this repo,
and since some of them are going to share structures and functions,
try to organize the repo in a more Go-like way, building
subpackages for the common code.

This might change slightly in the future, but for the moment it
feels good enough.