Add before_script for starting up services
We need to have RabbitMQ running with a certain set of configuration
if we want to be able to test properly the code.
Add the initial script and configuration.

This is ofc untested since the only way to test it AFAIK is to push
to the repo and wait for Travis-CI to pick up the commit and try
to build it.