First implementation of a working worker
This is just the first shot at implementing the worker.
It is capable of consuming messages from the specified queue, and
then simply logging them to the log file after unmarshaling
the object passed through the wire.

Still there is one major bug that has been introduced in the API
portion, and it's the fact that after receiving the first message
it doesn't process correctly any subsequent request.

The worker doesn't seem to be affected by the same issue, and
I suspect this has been caused by the refactoring of the whole
AMQP package.